+++ solstice/src/main/java/dev/equo/ide/BuildPluginIdeMain.java
@@ -99,6 +99,10 @@ public class BuildPluginIdeMain {
 			}
 			var vmArgs = new ArrayList<String>();
 			if (EquoChromium.isEnabled(ideHooks)) {
+				// This property is used to fix the error in the setUrl method.
+				vmArgs.add("-Dchromium.args=--disable-site-isolation-trials");
+				// This property improve loading time of setText for large resources.
+				vmArgs.add("-Dchromium.setTextAsUrl=file:");
 				Patch.patch(classpathSorted, nestedJarFolder, "patch-chromium-swt");
 				SignedJars.stripIf(classpathSorted, fileName -> fileName.startsWith("org.eclipse.swt."));
 				vmArgs.add(