+++ solstice/src/main/java/dev/equo/ide/Catalog.java
@@ -72,7 +72,7 @@ public class Catalog implements Comparable<Catalog> {
 	public static final Catalog GRADLE_BUILDSHIP =
 			new Catalog(
 					"gradleBuildship",
-					"https://download.eclipse.org/buildship/updates/e423/releases/3.x/" + V,
+					"https://download.eclipse.org/buildship/updates/e427/releases/3.x/" + V,
 					jre11("3.1.7.v20230428-1420"),
 					List.of("org.eclipse.buildship.feature.group"),
 					JDT);