+++ solstice/CHANGELOG.md
@@ -3,6 +3,8 @@
 We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format.
 
 ## [Unreleased]
+
+## [1.5.0] - 2023-06-22
 ### Added
 - `Catalog.PureMaven` for plugins which are only published to maven, not p2. ([#143](https://github.com/equodev/equo-ide/pull/143))
 - `P2Model.getPureMaven()` which has a set of maven coordinates which do not partipate in the p2 resolution at all. ([#143](https://github.com/equodev/equo-ide/pull/143))